Move class CuckooWatch into a separate file
