+++ src/artm/core/cuckoo_watch.cc
@@ -1,0 +1,9 @@
+// Copyright 2016, Additive Regularization of Topic Models.
+
+#include "artm/core/cuckoo_watch.h"
+
+namespace artm {
+namespace core {
+
+}  // namespace core
+}  // namespace artm
