+++ src/artm/core/dense_phi_matrix.h
@@ -7,6 +7,7 @@
 #include <map>
 #include <memory>
 #include <string>
+#include <unordered_map>
 #include <vector>
 
 #include "boost/utility.hpp"
