+++ src/artm/core/dictionary.h
@@ -6,6 +6,7 @@
 #include <map>
 #include <string>
 #include <vector>
+#include <unordered_map>
 #include <utility>
 
 #include "artm/messages.pb.h"
