+++ utils/cpplint_files.txt
@@ -5,6 +5,7 @@ src/artm/core/batch_manager.cc
 src/artm/core/cache_manager.cc
 src/artm/core/collection_parser.cc
 src/artm/core/dictionary.cc
+src/artm/core/cuckoo_watch.cc
 src/artm/core/dense_phi_matrix.cc
 src/artm/core/helpers.cc
 src/artm/core/instance.cc
@@ -59,6 +60,7 @@ src/artm/core/call_on_destruction.h
 src/artm/core/check_messages.h
 src/artm/core/collection_parser.h
 src/artm/core/common.h
+src/artm/core/cuckoo_watch.h
 src/artm/core/dense_phi_matrix.h
 src/artm/core/dictionary.h
 src/artm/core/exceptions.h
